requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
groq
streamlit
python-dotenv
tiktoken
